requirements.txt
flask==3.0.3
flask-socketio==5.3.6
gunicorn==22.0.0
python-socketio==5.11.2
gevent-websocket==0.10.1
